refactor(tests): enhance Jest setup and improve component mock implementations
- Updated `jest-setup.js` to streamline mock implementations for `ThemedText` and `react-native` components, improving clarity.
- Added error logging in `ErrorBoundary` to capture errors during rendering.
- Refactored `ChoicesSection` to simplify conditional rendering logic for result and explanation modals.
- Removed unnecessary `data-testid` attributes from `QuizChoice` and corresponding snapshots to reduce redundancy.
- Introduced `useQuizRecording` hook for recording quiz answers, enhancing functionality and database interaction.
- Updated tests for `useQuizRecording` to validate hook import and functionality without rendering, addressing previous timeout issues.
