+++ components/ErrorBoundary.tsx
@@ -25,6 +25,11 @@ export class ErrorBoundary extends Component<
     return { hasError: true, error }
   }
 
+  componentDidCatch(error: Error, errorInfo: React.ErrorInfo) {
+    // Log the error to an error reporting service
+    console.error('Error caught by ErrorBoundary:', error, errorInfo)
+  }
+
   render() {
     if (this.state.hasError) {
       return (
